refactor(api): move getGroup in bounded context
