+++ api/lib/application/organizations/index.js
@@ -600,24 +600,6 @@ const register = async function (server) {
         ],
       },
     },
-    {
-      method: 'GET',
-      path: '/api/organizations/{id}/groups',
-      config: {
-        pre: [{ method: securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents }],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-        },
-        handler: organizationController.getGroups,
-        tags: ['api', 'organizations'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés',
-          'Elle retourne les groupes rattachés à l’organisation.',
-        ],
-      },
-    },
   ];
 
   server.route([...adminRoutes, ...orgaRoutes]);
