+++ api/src/prescription/organization-learner/application/sup-learner-list-route.js
@@ -40,6 +40,24 @@ const register = async function (server) {
         ],
       },
     },
+    {
+      method: 'GET',
+      path: '/api/organizations/{id}/groups',
+      config: {
+        pre: [{ method: securityPreHandlers.checkUserBelongsToSupOrganizationAndManagesStudents }],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.organizationId,
+          }),
+        },
+        handler: supLearnerListController.getGroups,
+        tags: ['api', 'organizations'],
+        notes: [
+          'Cette route est restreinte aux utilisateurs authentifiés',
+          'Elle retourne les groupes rattachés à l’organisation.',
+        ],
+      },
+    },
   ]);
 };
