+++ api/tests/prescription/organization-learner/unit/domain/usecases/find-groups-by-organization_test.js
@@ -1,5 +1,5 @@
-import { findGroupsByOrganization } from '../../../../lib/domain/usecases/find-groups-by-organization.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { findGroupsByOrganization } from '../../../../../../src/prescription/organization-learner/domain/usecases/find-groups-by-organization.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | find-groups-by-organization', function () {
   it('should return all groups', async function () {
